slide-panel-mannager, slide-panel, theme-provider
