+++ packages/terra-slide-panel-manager/src/terra-dev-site/doc/slide-panel-manager/SlidePanelManager.1.doc.jsx
@@ -1,15 +1,15 @@
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-duplicates */
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import { name } from '../../../package.json';
-import ReadMe from '../../../docs/README.md';
+import { name } from '../../../../package.json';
+import ReadMe from '../../../../docs/README.md';
 
 // Component Source
-import SlidePanelManagerSrc from '!raw-loader!../../../src/SlidePanelManager.jsx';
+import SlidePanelManagerSrc from '!raw-loader!../../../../src/SlidePanelManager.jsx';
 
 // Examples
-import SlidePanelManagerExample from './example/SlidePanelManagerExample';
-import SlidePanelManagerExampleSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SlidePanelManagerExample.jsx';
+import SlidePanelManagerExample from '../example/SlidePanelManagerExample';
+import SlidePanelManagerExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SlidePanelManagerExample.jsx';
 
 const DocPage = () => (
   <DocTemplate
