+++ packages/terra-slide-panel/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Change logs to documentation on terra-dev-site
 
 2.9.0 - (June 12, 2018)
 ------------------
