+++ packages/terra-slide-panel/src/terra-dev-site/doc/slide-panel/ChangeLog.2.doc.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DocTemplate from 'terra-doc-template';
+import ChangeLog from '../../../../CHANGELOG.md';
+
+const DocPage = () => (
+  <DocTemplate
+    packageName="terra-slide-panel"
+    srcPath="https://github.com/cerner/terra-framework/tree/master/packages/terra-slide-panel"
+    readme={ChangeLog}
+  />
+);
+
+export default DocPage;
