+++ packages/terra-theme-provider/src/terra-dev-site/doc/theme-provider/ChangeLog.2.doc.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DocTemplate from 'terra-doc-template';
+import ChangeLog from '../../../../CHANGELOG.md';
+
+const DocPage = () => (
+  <DocTemplate
+    packageName="terra-theme-provider"
+    srcPath="https://github.com/cerner/terra-framework/tree/master/packages/terra-theme-provider"
+    readme={ChangeLog}
+  />
+);
+
+export default DocPage;
