+++ packages/terra-theme-provider/src/terra-dev-site/doc/theme-provider/ThemeProvider.1.doc.jsx
@@ -1,15 +1,15 @@
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-duplicates */
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import { name } from '../../../package.json';
-import ReadMe from '../../../docs/README.md';
+import { name } from '../../../../package.json';
+import ReadMe from '../../../../docs/README.md';
 
 // Component Source
-import ThemeProviderSrc from '!raw-loader!../../../src/ThemeProvider';
+import ThemeProviderSrc from '!raw-loader!../../../../src/ThemeProvider';
 
 // Example Files
-import DefaultThemeProvider from './example/DefaultThemeProvider';
-import DefaultThemeProviderSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/DefaultThemeProvider.jsx';
+import DefaultThemeProvider from '../example/DefaultThemeProvider';
+import DefaultThemeProviderSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DefaultThemeProvider.jsx';
 
 const DocPage = () => (
   <DocTemplate
